backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
google-generativeai>=0.3.1
chromadb>=0.4.18
PyPDF2>=3.0.1
python-pptx>=0.6.23
sentence-transformers>=2.2.2
pydantic>=2.5.0
pydantic-settings>=2.1.0
pytest>=7.4.0
httpx>=0.25.0
